refactor(tarko): clean up agent server exclusive mode implementation
Remove unused statusBroadcastInterval and simplify exclusive mode logic
Replace any types with proper TypeScript types in tests
