+++ multimodal/tarko/agent-server/vitest.config.mts
@@ -3,10 +3,20 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 import { defineConfig } from 'vitest/config';
+import path from 'path';
 
 export default defineConfig({
   test: {
-    environment: 'node',
     include: ['**/*.test.ts'],
+    environment: 'node',
+    globals: true,
+    setupFiles: ['./tests/setup.ts'],
+    testTimeout: 10000,
+    hookTimeout: 10000,
+  },
+  resolve: {
+    alias: {
+      '@': path.resolve(__dirname, './src'),
+    },
   },
 });
